Speed-up PyCharm parsing of remote libraries
PyCharm scan all the paths in $PYTHONPATH, looking for python libraries to parse for auto-completion. It also looks for .so file in case of it's a compiled library.

When doing a cmsenv, the CMSSW/lib/ folders are also added to the $PYTHONPATH. As a result, PyCharm takes ages to do its initial parsing because it has to scan for hundred of libraries which we know for sure are **not** python library.

This commit removes the CMSSW/lib paths from PYTHONPATH (both for local release as well as base release). Initial parsing time for PyCharm is reduced from about 1 hour to 1 minute.